break out opening file
