+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudwatch/logs/actions/OpenCurrentInEditor.kt
@@ -6,22 +6,14 @@ package software.aws.toolkits.jetbrains.services.cloudwatch.logs.actions
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.application.ModalityState
-import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.fileTypes.PlainTextLanguage
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiFileFactory
 import com.intellij.util.ui.ListTableModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import software.amazon.awssdk.services.cloudwatchlogs.model.OutputLogEvent
 import software.aws.toolkits.jetbrains.utils.getCoroutineUiContext
-import software.aws.toolkits.jetbrains.utils.notifyError
 import software.aws.toolkits.resources.message
 
 class OpenCurrentInEditor(
@@ -30,7 +22,7 @@ class OpenCurrentInEditor(
     private val logsTableModel: ListTableModel<OutputLogEvent>
 ) :
     AnAction(message("cloudwatch.logs.open_in_editor"), null, AllIcons.Actions.Menu_open), CoroutineScope by GlobalScope, DumbAware {
-    private val edt = getCoroutineUiContext(ModalityState.defaultModalityState())
+    private val edt = getCoroutineUiContext()
 
     override fun actionPerformed(e: AnActionEvent) {
         launch {
@@ -39,25 +31,7 @@ class OpenCurrentInEditor(
     }
 
     private suspend fun actionPerformedSuspend(e: AnActionEvent) {
-        val factory = PsiFileFactory.getInstance(project)
-        val file: PsiFile = factory.createFileFromText(
-            logStream,
-            PlainTextLanguage.INSTANCE,
-            logsTableModel.items.joinToString("") { if (it.message().endsWith("\n")) it.message() else "${it.message()}\n" },
-            true,
-            false,
-            true
-        )
-        withContext(edt) {
-            file.virtualFile?.let {
-                ApplicationManager.getApplication().runWriteAction {
-                    it.isWritable = false
-                }
-                // set virtual file to read only
-                FileEditorManager.getInstance(project).openFile(it, true, true).ifEmpty {
-                    notifyError(message("cloudwatch.logs.open_in_editor_failed"))
-                }
-            }
-        }
+        val fileContent = logsTableModel.items.joinToString("") { if (it.message().endsWith("\n")) it.message() else "${it.message()}\n" }
+        OpenStreamInEditor.open(project, edt, logStream, fileContent)
     }
 }
