+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudwatch/logs/actions/OpenLogStreamInEditor.kt
@@ -6,23 +6,16 @@ package software.aws.toolkits.jetbrains.services.cloudwatch.logs.actions
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ModalityState
-import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.fileTypes.PlainTextLanguage
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiFileFactory
 import com.intellij.ui.table.JBTable
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient
 import software.aws.toolkits.jetbrains.core.AwsClientManager
 import software.aws.toolkits.jetbrains.utils.ApplicationThreadPoolScope
 import software.aws.toolkits.jetbrains.utils.getCoroutineUiContext
-import software.aws.toolkits.jetbrains.utils.notifyError
 import software.aws.toolkits.resources.message
 
 class OpenLogStreamInEditor(
@@ -46,25 +39,7 @@ class OpenLogStreamInEditor(
         val row = groupTable.selectedRow.takeIf { it >= 0 } ?: return
         val logStream = groupTable.getValueAt(row, 0) as String
         val events = client.getLogEventsPaginator { it.startFromHead(true).logGroupName(logGroup).logStreamName(logStream) }
-        val factory = PsiFileFactory.getInstance(project)
-        val file: PsiFile = factory.createFileFromText(
-            logStream,
-            PlainTextLanguage.INSTANCE,
-            events.events().filterNotNull().joinToString("") { if (it.message().endsWith("\n")) it.message() else "${it.message()}\n" },
-            true,
-            false,
-            true
-        )
-        withContext(edt) {
-            file.virtualFile?.let {
-                ApplicationManager.getApplication().runWriteAction {
-                    it.isWritable = false
-                }
-                // set virtual file to read only
-                FileEditorManager.getInstance(project).openFile(it, true, true).ifEmpty {
-                    notifyError(message("cloudwatch.logs.open_in_editor_failed"))
-                }
-            }
-        }
+        val fileContent = events.events().filterNotNull().joinToString("") { if (it.message().endsWith("\n")) it.message() else "${it.message()}\n" }
+        OpenStreamInEditor.open(project, edt, logStream, fileContent)
     }
 }
